requirements.txt
streamlit
streamlit-mic-recorder
google-genai
google-cloud-aiplatform
gTTS
SpeechRecognition
pandas